requirements.txt
# --- Core App ---
streamlit>=1.32.0
python-dotenv>=1.0.0

# --- LLM ---
google-generativeai>=0.5.0

# --- Agent Orchestration ---
langgraph>=0.0.40

# --- Vector DB ---
chromadb>=0.4.24

# --- Embeddings ---
sentence-transformers>=2.2.2
torch>=2.0.0

# --- Utilities ---
typing-extensions>=4.7.0

# --- Testing ---
pytest>=7.4.0
